폴더 코드 정리
- util.ts :: 함수 역할에 따라 폴더분리
- 사용하지 않는 컴포넌트 삭제
