+++ src/components/organisms/Community/AdminCommunityDetail.tsx
@@ -2,7 +2,7 @@ import { useAuthStore } from '@/@features/Auth/useAuthStore';
 import { CommunityDetailRes } from '@/@features/Community/type';
 import ToastViewer from '@/components/atoms/ToastViewer';
 import * as S from '@/styles/index.style';
-import { getDateString } from '@/util/util';
+import { getDateString } from '@/utils/stringUtil';
 import { ReactNode } from 'react';
 
 interface Props {
